+++ authpass/CHANGELOG.md
@@ -3,6 +3,7 @@
 * One Drive: Add conflict detection.
 * Kdbx: Improved compatibility with other Keepass apps.
 * Password generator: Allow generted passwords of custom length >40 characters.
+* WebDAV: Improved conflict detection on some webdav servers.
 
 # 1.7.5